chore: respect configured entrypoint in application compiler
